Move functions where they belong
